requirements.txt
Django==2.2.6
mysql-connector-python==8.0.18
django-mptt==0.10.0
django-mptt-admin==0.7.2
